# Conflicts:
#	InventoryAdminUi/Test/Mftf/ActionGroup/AdminAssertAssignedSourcesQtyInGridForProductActionGroup.xml
#	InventoryAdminUi/Test/Mftf/ActionGroup/AdminBulkUnAssignSourcesToBulkSimpleProductsActionGroup.xml
#	InventoryCatalog/etc/queue_publisher.xml
#	InventoryExportStockApi/Test/Api/ExportStockSalableQtyTest.php
#	InventoryInStorePickupAdminUi/Test/Mftf/Section/AdminEditSourcePickupLocationSection.xml

+++ InventoryShipping/Model/GetItemsToDeductFromShipment.php
@@ -189,7 +189,7 @@ class GetItemsToDeductFromShipment
     private function castQty(OrderItem $item, $qty)
     {
         if ($item->getIsQtyDecimal()) {
-            $qty = (double)$qty;
+            $qty = (float)$qty;
         } else {
             $qty = (int)$qty;
         }
